requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-google-vertexai
google-cloud-aiplatform
unstructured
requests
beautifulsoup4
pandas
openpyxl
faiss-cpu
sentence-transformers

langchain-huggingface
